refactor: mv src/gui/src/ServerConfigDialog => src/gui/src/dialogs/ServerConfigDialog
