+++ src/gui/src/dialogs/ServerConfigDialog.ui
